(BSR)[BO] refactor: clean types import for offerers & venues routes
